Create Image.py
